Move multi section functions to utils. Merge ones unique to a class into the class. Update docs and examples
