+++ openaerostruct/docs/advanced_features/scripts/basic_2_sec_construction.py
@@ -6,9 +6,9 @@ import numpy as np
 import openmdao.api as om
 from openaerostruct.geometry.geometry_group import MultiSecGeometry
 from openaerostruct.aerodynamics.aero_groups import AeroPoint
-from openaerostruct.geometry.geometry_group import build_sections
-from openaerostruct.geometry.geometry_unification import unify_mesh
-from openaerostruct.geometry.multi_unified_bspline_utils import build_multi_spline, connect_multi_spline
+from openaerostruct.geometry.utils import build_section_dicts
+from openaerostruct.geometry.utils import unify_mesh
+from openaerostruct.geometry.utils import build_multi_spline, connect_multi_spline
 import matplotlib.pyplot as plt
 
 # docs checkpoint 1
@@ -112,9 +112,9 @@ control points at each section junction to an index of a global control vector. 
 # In order to construct this global B-spline control vector we first need to generate the unified surface mesh.
 # The unified surface mesh is simply all the individual section surface meshes combine into a single unified OAS mesh array.
 
-# First we will call the utility function build_sections which takes the surface dictionary and outputs a list of surface dictionaries corresponding to
+# First we will call the utility function build_section_dicts which takes the surface dictionary and outputs a list of surface dictionaries corresponding to
 # each section.
-section_surfaces = build_sections(surface)
+section_surfaces = build_section_dicts(surface)
 
 # We can then call unify_mesh which outputs the unified mesh of all of the sections.
 uniMesh = unify_mesh(section_surfaces)
