+++ tests/integration_tests/test_multi_3_sec.py
@@ -10,8 +10,8 @@ class Test(unittest.TestCase):
 
         from openaerostruct.geometry.geometry_group import MultiSecGeometry
         from openaerostruct.aerodynamics.aero_groups import AeroPoint
-        from openaerostruct.geometry.geometry_group import build_sections
-        from openaerostruct.geometry.geometry_unification import unify_mesh
+        from openaerostruct.geometry.utils import build_section_dicts
+        from openaerostruct.geometry.utils import unify_mesh
         from openaerostruct.utils.testing import get_three_section_surface
 
         # Create a dictionary with info and options about the multi-section aerodynamic
@@ -45,7 +45,7 @@ class Test(unittest.TestCase):
 
         # Generate the sections and unified mesh here in addition to adding the components.
         # This has to ALSO be done here since AeroPoint has to know the unified mesh size.
-        section_surfaces = build_sections(surface)
+        section_surfaces = build_section_dicts(surface)
         uni_mesh = unify_mesh(section_surfaces)
         surface["mesh"] = uni_mesh
 
@@ -113,8 +113,8 @@ class Test(unittest.TestCase):
 
         from openaerostruct.geometry.geometry_group import MultiSecGeometry
         from openaerostruct.aerodynamics.aero_groups import AeroPoint
-        from openaerostruct.geometry.geometry_group import build_sections
-        from openaerostruct.geometry.geometry_unification import unify_mesh
+        from openaerostruct.geometry.utils import build_section_dicts
+        from openaerostruct.geometry.utils import unify_mesh
         from openaerostruct.utils.testing import get_three_section_surface
 
         surface, sec_chord_cp = get_three_section_surface(sym=False, visc=False)
@@ -146,7 +146,7 @@ class Test(unittest.TestCase):
 
         # Generate the sections and unified mesh here in addition to adding the components.
         # This has to ALSO be done here since AeroPoint has to know the unified mesh size.
-        section_surfaces = build_sections(surface)
+        section_surfaces = build_section_dicts(surface)
         uni_mesh = unify_mesh(section_surfaces)
         surface["mesh"] = uni_mesh
