+++ tests/multisection_tests/test_joining_comp.py
@@ -2,14 +2,14 @@ import numpy as np
 import unittest
 
 from openaerostruct.geometry.geometry_multi_join import GeomMultiJoin
-from openaerostruct.geometry.geometry_group import build_sections
+from openaerostruct.geometry.utils import build_section_dicts
 from openaerostruct.utils.testing import run_test, get_three_section_surface
 
 
 class Test(unittest.TestCase):
     def test(self):
         (surface, chord_bspline) = get_three_section_surface()
-        sec_dicts = build_sections(surface)
+        sec_dicts = build_section_dicts(surface)
 
         comp = GeomMultiJoin(sections=sec_dicts, dim_constr=[np.ones(3), np.ones(3), np.ones(3)])
